add one awesome operation!
self.mod + button4 (scroll up):  to minimize all the other clients with
current screen
self.mod + button5 (scroll down): to restore all the clients with
current screen

While using the transparent plus the awesome desktop widgets, it is
quite annoying while the transparent clients such as terminal on the
top of another client.

This is a counterpart of "self.mod + TAB", cause it minimized all
the other clients.